+++ src/Command/RunCommand.php
@@ -3,10 +3,12 @@
 namespace Orisai\Scheduler\Command;
 
 use Generator;
+use Orisai\Clock\SystemClock;
 use Orisai\Scheduler\Scheduler;
 use Orisai\Scheduler\Status\JobResultState;
 use Orisai\Scheduler\Status\JobSummary;
 use Orisai\Scheduler\Status\RunSummary;
+use Psr\Clock\ClockInterface;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -19,9 +21,9 @@ final class RunCommand extends BaseRunCommand
 
 	private Scheduler $scheduler;
 
-	public function __construct(Scheduler $scheduler)
+	public function __construct(Scheduler $scheduler, ?ClockInterface $clock = null)
 	{
-		parent::__construct();
+		parent::__construct($clock ?? new SystemClock());
 		$this->scheduler = $scheduler;
 	}