+++ CHANGELOG.md
@@ -30,6 +30,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `ListCommand`
 	- on invalid input option - writes error to output instead of throwing exception
+- `JobInfo`
+	- `getStart()` includes correct timezone (in main process, when using `ProcessJobExecutor`)
+- `JobResult`
+	- `getEnd()` includes correct timezone (in main process, when using `ProcessJobExecutor`)
 
 ## [2.0.0](https://github.com/orisai/scheduler/compare/1.0.0...2.0.0) - 2024-01-26